correctly handle paragraphs inside questions
This fixes the rendering of the question introduced in the previous
patch, but also adds <p>...</p> around all other questions.

This is done by wrapping the body of question environments in an
<ltx:inline-para>...</ltx:inline-para> block.  LaTeXML will detect
this an properly split paragraphs inside those.  (I learned the hard
way that <ltx:para> is not allowed in <ltx:note>.)

This adds another nesting level to answers, which usually appear in
the last paragraph of the question, so multiple XPATH selectors had to
be adjusted to retrieve answers at arbitrary depth.